requirements.txt
aiogram
python-dotenv
langchain
langchain-community
openai
faiss-cpu
PyPDF2
tiktoken
